# Implement Stream-Safe Text Process and QuickCheck algorithms