… Cargo.toml
… scripts/unicode_gen_normtests.py
… src/normalization_tests.rs
… src/quick_check.rs
… src/recompose.rs
… src/stream_safe.rs
… src/tables.rs
… src/test.rs
… src/testdata.rs